feat!: implement traits on encounter generator
