+++ src/db/db_proxy.rs
@@ -73,6 +73,13 @@ fn fetch_creatures_passing_single_filter(
                 filter_vec.contains(creature.clone().family.unwrap_or_default().as_str())
             })
             .collect(),
+        CreatureFilter::Traits => cr_iterator
+            .filter(|creature| {
+                filter_vec
+                    .iter()
+                    .any(|curr_trait| creature.clone().traits.contains(curr_trait))
+            })
+            .collect(),
         CreatureFilter::Alignment => cr_iterator
             .filter(|creature| filter_vec.contains(creature.alignment.to_string().as_str()))
             .collect(),
